requirements.txt
streamlit
pandas
numpy
scikit-learn
fpdf2
matplotlib